Move legacy bookmark copy use case to feature application module
