+++ web/src/routes/(app)/[slug=npub]/bookmarks/+page.svelte
@@ -24,7 +24,7 @@
 		resolveSelectedBookmarkList
 	} from './BookmarkListTabs';
 	import { BookmarkPageState } from './BookmarkPageState.svelte';
-	import { copyLegacyBookmarks } from '$lib/author/BookmarkCopy';
+	import { copyLegacyBookmarks } from '$lib/features/bookmarks/application/copy-legacy-bookmarks';
 	import { addToast } from '$lib/components/Toaster.svelte';
 	import { deleteLegacyBookmarks } from '$lib/author/legacy-bookmark-delete';
